Move v1 response handling to V1Context
The V1-specific error handling was being done in PsbtContext even though
the PsbtContext should not know about V1-related json errors.
